Refactor Customer to Partner and PartnerMapper to PatientMapper
