+++ fhir-odoo-mapper/src/main/java/com/ozonehis/fhir/odoo/mappers/SaleOrderMapper.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.ozonehis.fhir.odoo.mappers.odoo;
+package com.ozonehis.fhir.odoo.mappers;
 
 import com.ozonehis.fhir.odoo.OdooConstants;
 import com.ozonehis.fhir.odoo.model.OdooResource;
